// frontend/src/generated/model/vocabularyListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { VocabularyListResponseQueryDefinition } from "./vocabularyListResponseQueryDefinition";

export interface VocabularyListResponse {
  color: string | null;
  list_id: string;
  list_type: string;
  member_sense_ids: string[];
  name: string;
  ordered: boolean;
  profile_id: string;
  purpose: string;
  query_definition: VocabularyListResponseQueryDefinition;
  revision_id: string;
  revision_no: number;
  status: string;
  tags: string[];
  variety_id: string;
  version: number;
}